requirements.txt
fastapi
uvicorn
transformers
torch
numpy
opencv-python-headless
pillow
pytesseract
